Server/functions/src/pythonScript/requirements.txt
firebase_functions~=0.1.0
Pillow==10.1.0
face_recognition==1.3.0
